Initial commit
just made a quick function that takes some inital x and a number of times to be run and creates an array of size N that stores values of H or (x^2). This isn't what we need it to do just a starting point. Next step is to calculate delta H and work in an if statement to accept/reject the new values depending on the probability